`icinga-stack`: Add chart for `Icinga Kubernetes`
Note that the chart is subject to change as `Icinga Kubernetes` is not
yet released, especially regarding the configuration. At the moment, a
configuration file is mounted from a `ConfigMap` that does not support
secrets, as `Icinga Kubernetes` lacks support for environment variables.
This will change in the near future.