`icinga-stack`: `Always` pull image for `Icinga Web`
Although the `Icinga Web` image tags are coupled to release tags,
several modules are included. We rebuild the image using the same latest
release tag when modules get new release, so the image needs to be
pulled again in such cases. Updates to the base image (OS) are also
pushed using the same tags. This does not only apply to this image, so
the pull policy for all other images should also be reconsidered.